moved `tryOrPrintIndexer` to `TestCommon`
